Get joint model group
The joint model group used to be stored in the parent class MoveItBase.
As this is not the case anymore, a parameter was added to choose the correct
joint model group and then robot_model_ is used to get the joint model group.

Additionally the variable jmg_ was removed as there is an identical variable
arm_jmg_. Occurences of jmg_ in the code were replaced by arm_jmg_.